pyproject.toml
[tool.poetry]
name = "telegatagger"
version = "0.1.0"
description = ""
authors = ["trojvn <mrx.trojan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
mongowrapper = {git = "https://github.com/trojvn/mongowrapper.git"}
rich = "^13.8.0"
rich-menu = "^0.2.0"
httpx = "^0.27.2"
telethon = "^1.36.0"
tooler = {git = "https://github.com/trojvn/tooler.git"}
jsoner = {git = "https://github.com/trojvn/jsoner.git"}
montywrapper = {git = "https://github.com/trojvn/montywrapper.git"}
python-socks = "2.1.1"
async-timeout = "^4.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
